+++ spec/opal/filters/unsupported/encoding.rb
@@ -48,3 +48,9 @@ opal_filter "Encoding" do
   fails "Magic comment is case-insensitive"
   fails "Magic comment determines __ENCODING__"
 end
+
+opal_filter "Regexp.escape" do
+  fails "Regexp.escape sets the encoding of the result to US-ASCII if there are only US-ASCII characters present in the input String"
+  fails "Regexp.escape sets the encoding of the result to the encoding of the String if any non-US-ASCII characters are present in an input String with valid encoding"
+  fails "Regexp.escape sets the encoding of the result to ASCII-8BIT if any non-US-ASCII characters are present in an input String with invalid encoding"
+end