+++ CHANGELOG.md
@@ -115,6 +115,8 @@
 
 *   Fix some stdlib `Date` methods.
 
+*   Fix `Regexp.escape` to properly escape \n, \t, \r, \f characters.
+
 ## 0.5.5 2013-11-25
 
 *   Fix regression: add `%i[foo bar]` style words back to lexer